Moved GUI.py to parent directory and deleted HTML
